Update publishing workflow, remove pyrokinetics dependency
Currently the pyrokinetics dependency works as the published
library does not depend on this package. However, due to the
circular dependencies, once the published version of
pyrokinetics depends on this library, all CI jobs in this
repo will begin to fail, as the version required by
pyrokinetics will never match the current development
version.

The solution is to remove the pyrokinetics dependency from
this library. It isn't necessary, as due to the use of
entry-points, the pyrokinetics-dependent files will only
ever be imported if the user's environment contains
pyrokinetics anyway.